Many Changes:
imported IOException handling

changed name from clientSocket for the socket to connection

tuned it for the GUI a bit

Formatting

Takes care of closing the socket once the client requests leave.

Added the sendMessage method which does not work. This will probably be
changed to somewhere in the while loop.

--Does not work at the moment--